\name{getMachineName}
\alias{getMachineName}
\alias{getMovieName}
\title{
  Get machine/movie names.
}
\description{
  Retrieve movie/machine name from cmp.h5 file.
}
\usage{
  getMachineName(cmpH5, idx = seq.int(1, nrow(cmpH5)))
}
\arguments{
  \item{cmpH5}{
  }
  \item{idx}{
  }
}
\keyword{classes}

