\name{getNumPasses}
\alias{getNumPasses}
\title{
  Returns the number of times a polymerase sequenced the insert.
}
\description{
  The number of passes is a whole number which approximates the number
  of times that the polymerase sequences the insert.
}
\usage{
  getNumPasses(cmpH5, idx = 1:nrow(cmpH5))
}
\arguments{
  \item{cmpH5}{
  }
  \item{idx}{
  }
}
\details{
  This function is only relevant for CCS-based PacBioCmpH5 files.
}
\keyword{classes}

