\name{getRegionsTable}
\alias{getRegionsTable}
\alias{getAcqParams}
\alias{getChannelToBaseMap}
\alias{getRegionsTable,MultiPart-method}
\alias{getAcqParams,MultiPart-method}
\alias{getChannelToBaseMap,MultiPart-method}
\alias{getRegionsTable,PacBioBasH5-method}
\alias{getAcqParams,PacBioBasH5-method}
\alias{getChannelToBaseMap,PacBioBasH5-method}
\title{
  High-level pulse/base file information.
}
\description{
  Return aspects of the movie acquisition.
}
\usage{
  getRegionsTable(h5Obj, ...)
}
\arguments{
  \item{h5Obj}{
    Either a PacBioBasH5 or PacBioPlsH5 file.
  }
  \item{...}{
    Currently ignored
    }
}
\keyword{classes}
