\name{getZMWs}
\alias{getZMWs}
\title{
  Get ZMWs from aln.h5 Files
}
\description{
  Retrieve the ZMW object which describes how things are stored in
  aln.h5 files.
}
\usage{
  getZMWs(alnH5)
}
\arguments{
  \item{alnH5}{
    An object of class \code{PacBioAlnH5}.
  }
}
\value{
  A matrix with various ZMW information for the aln.h5 files.
}
\keyword{datasets}
