\name{pbh5-internal}
\title{Internal pbh5 objects}
\alias{.bMap}
\alias{.bMapC}
\alias{.f}
\alias{.getAlignmentsWithFeaturesAsDataFrame}
\alias{.hasDataset}
\alias{.matchVersion}
\alias{.NUMBERS}
\alias{.ofBasecalls}
\alias{.ofPulses}
\alias{.PAIRS}
\description{Internal pbh5 objects.}
\details{These are not to be called by the user.}
\keyword{internal}